# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.NodeInstallationRequest:
    properties:
      api_key:
        type: string
      max_storage:
        minimum: 1
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      port:
        maximum: 65535
        minimum: 1
        type: integer
      storage_path:
        type: string
    required:
    - name
    - storage_path
    type: object
  models.NodeInstallationResponse:
    properties:
      config_path:
        type: string
      install_path:
        type: string
      message:
        type: string
      node:
        $ref: '#/definitions/models.StorageNodeResponse'
      start_command:
        type: string
      success:
        type: boolean
    type: object
  models.NodeSetupRequest:
    properties:
      master_api_key:
        type: string
      master_url:
        type: string
      max_storage:
        minimum: 1
        type: integer
      node_api_key:
        minLength: 10
        type: string
      node_name:
        maxLength: 100
        minLength: 3
        type: string
      storage_path:
        type: string
    required:
    - master_api_key
    - master_url
    - node_api_key
    - node_name
    - storage_path
    type: object
  models.RegisterNodeRequest:
    properties:
      auth_key:
        minLength: 32
        type: string
      is_active:
        type: boolean
      max_storage:
        minimum: 0
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      priority:
        maximum: 100
        minimum: 0
        type: integer
      url:
        type: string
    required:
    - auth_key
    - name
    - url
    type: object
  models.StorageNodeResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_healthy:
        type: boolean
      last_ping:
        type: string
      max_storage:
        type: integer
      name:
        type: string
      priority:
        type: integer
      updated_at:
        type: string
      url:
        type: string
      used_storage:
        type: integer
    type: object
...
  node.ListNodesResponse:
    properties:
      limit:
        type: integer
      message:
        type: string
      nodes:
        items:
          $ref: '#/definitions/models.StorageNodeResponse'
        type: array
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
  node.RegisterNodeResponse:
    properties:
      message:
        type: string
      node:
        $ref: '#/definitions/models.StorageNodeResponse'
      success:
        type: boolean
    type: object
...
  setup.NodeSetupResponse:
    properties:
      config:
        additionalProperties: true
        type: object
      message:
        type: string
      node:
        $ref: '#/definitions/models.StorageNodeResponse'
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@shbucket.local
    name: SHBucket Support
  description: SHBucket is a distributed object storage system similar to AWS S3
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SHBucket API
  version: 2.0.0
paths:
...
  /node/auth-key:
    get:
      consumes:
      - application/json
      description: Retrieve the authentication key for a specific node by URL
      parameters:
      - description: Node URL
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node authentication key
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Node not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get node auth key
      tags:
      - nodes
...
  /node/register:
    post:
      consumes:
      - application/json
      description: Allow a node to register itself without authentication (for node
        setup)
      parameters:
      - description: Node self-registration details
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Node registered successfully with auth key
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Self-register storage node
      tags:
      - nodes
  /nodes:
    get:
      consumes:
      - application/json
      description: Get a list of all storage nodes in the distributed system
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: false
        description: Show only active nodes
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Nodes retrieved successfully
          schema:
            $ref: '#/definitions/node.ListNodesResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List storage nodes
      tags:
      - nodes
    post:
      consumes:
      - application/json
      description: Register a new storage node in the distributed system
      parameters:
      - description: Node registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterNodeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Node registered successfully
          schema:
            $ref: '#/definitions/node.RegisterNodeResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Register storage node
      tags:
      - nodes
  /nodes/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a storage node from the distributed system
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Node not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete storage node
      tags:
      - nodes
...
  /nodes/install:
    post:
      consumes:
      - application/json
      description: Install and configure a new storage node
      parameters:
      - description: Node installation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.NodeInstallationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Node installed successfully
          schema:
            $ref: '#/definitions/models.NodeInstallationResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Install storage node
      tags:
      - nodes
...
  /setup/node:
    post:
      consumes:
      - application/json
      description: Initialize the system as a storage node connected to a master
      parameters:
      - description: Node setup configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.NodeSetupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Node setup successful
          schema:
            $ref: '#/definitions/setup.NodeSetupResponse'
        "400":
          description: Invalid request or setup failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Setup node
      tags:
      - setup
...
securityDefinitions:
  ApiKeyAuth:
    description: API Key for authentication
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"